\name{Ruxin-package}
\alias{Ruxin-package}
\docType{package}
\title{
  This package provides several functions for ranking numbers from small to large.
}
\description{
  R codes for ranking numbers are provided in this package
}
\author{
Ruxin Shi

Maintainer: Ruxin Shi <shiruxin@msu.edu>
}
\keyword{ sort }
